requirements.txt
aiogram==3.18.0
requests==2.32.3
httpx==0.28.1
openai==1.64.0
gspread==6.1.4
oauth2client==4.1.3
pytest==8.3.4
python-dotenv==1.0.1
pip==25.0.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.38
alembic==1.14.1
